개발 리뉴얼

+++ src/main/kotlin/com/youdozi/demo/repository/ArticleRepository.kt
@@ -3,7 +3,7 @@ package com.youdozi.demo.repository
 import com.youdozi.demo.entity.Article
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ArticleRepository : JpaRepository<Article, Int> {
+interface ArticleRepository : JpaRepository<Article, Long> {
 
     fun findByName(name : String) : Article?
 }
